Removed metadata and initializationActions fields from GoogleCloudConfig; removed jar and zip fields from HailConfig; CloudSdkDataProcClient => CloudSdkDataProcWrapper; removed startCluster() from CloudSdkDataProcWrapper
