+++ src/main/scala/loamstream/googlecloud/GoogleCloudConfig.scala
@@ -25,7 +25,6 @@ final case class GoogleCloudConfig(
     projectId: String,
     clusterId: String,
     credentialsFile: Path,
-    metadata: Option[String] = Defaults.metadata,
     zone: String = Defaults.zone,
     masterMachineType: String = Defaults.masterMachineType,
     masterBootDiskSize: Int = Defaults.masterBootDiskSize,
@@ -37,12 +36,10 @@ final case class GoogleCloudConfig(
     imageVersion: String = Defaults.imageVersion,
     scopes: String = Defaults.scopes,
     properties: String = Defaults.properties,
-    initializationActions: String = Defaults.initializationActions,
     maxClusterIdleTime: String = Defaults.maxClusterIdleTime)
 
 object GoogleCloudConfig extends Loggable {
   object Defaults { // for creating a minimal cluster
-    val metadata: Option[String] = None
     val zone: String = "us-central1-b"
     val masterMachineType: String = "n1-standard-1"
     val masterBootDiskSize: Int = 20 // in GB
@@ -58,7 +55,6 @@ object GoogleCloudConfig extends Loggable {
       "spark:spark.driver.memory=45g,spark:spark.driver.maxResultSize=30g,spark:spark.task.maxFailures=20," +
       "spark:spark.kryoserializer.buffer.max=1g,hdfs:dfs.replication=1"
     }
-    val initializationActions: String = "gs://loamstream/hail/hail-init.sh"
     val maxClusterIdleTime: String = "10m"
   }
   
@@ -68,7 +64,6 @@ object GoogleCloudConfig extends Loggable {
       projectId: String,
       clusterId: String,
       credentialsFile: Path,
-      metadata: Option[String] = Defaults.metadata,
       zone: String = Defaults.zone,
       masterMachineType: String = Defaults.masterMachineType,
       masterBootDiskSize: Int = Defaults.masterBootDiskSize,
@@ -80,7 +75,6 @@ object GoogleCloudConfig extends Loggable {
       imageVersion: String = Defaults.imageVersion,
       scopes: String = Defaults.scopes,
       properties: String = Defaults.properties,
-      initializationActions: String = Defaults.initializationActions,
       maxClusterIdleTime: String = Defaults.maxClusterIdleTime) {
     
     def toGoogleCloudConfig: Try[GoogleCloudConfig] = {
@@ -93,7 +87,6 @@ object GoogleCloudConfig extends Loggable {
           projectId = projectId,
           clusterId = clusterId,
           credentialsFile = credentialsFile,
-          metadata = metadata,
           zone = zone,
           masterMachineType = masterMachineType,
           masterBootDiskSize = masterBootDiskSize,
@@ -105,7 +98,6 @@ object GoogleCloudConfig extends Loggable {
           imageVersion = imageVersion,
           scopes = scopes,
           properties = properties,
-          initializationActions = initializationActions,
           maxClusterIdleTime = maxClusterIdleTime)
       }
     }
