+++ src/main/scala/loamstream/googlecloud/HailConfig.scala
@@ -18,12 +18,8 @@ import loamstream.util.Paths
  * Feb 22, 2017
  */
 final case class HailConfig(
-    jar: URI, 
-    zip: URI,
     condaEnv: String,
     scriptDir: Path = Defaults.scriptDir) extends HasScriptDir {
-  
-  def jarFile: String = java.nio.file.Paths.get(jar.getPath).getFileName.toString
   
   def withScriptDir(newScriptDir: Path): HailConfig = copy(scriptDir = newScriptDir)
 }
@@ -34,11 +30,10 @@ object HailConfig {
   }
   
   def fromConfig(config: Config): Try[HailConfig] = {
-    //NB: Import all Ficus typeclasses except the one for reading java.net.URIs, since we want our own handling of 
-    //those by ValueReaders.GcsUriReader.
+    //NB: Import all Ficus typeclasses except the one for reading java.net.URIs, since without doing this, automatic 
+    //unmarshalling fails.  (I'm not sure why :\ -Clint Jul 22, 2019) 
     import net.ceedubs.ficus.Ficus.{javaURIReader => _, _}
     import net.ceedubs.ficus.readers.ArbitraryTypeReader._
-    import ValueReaders.GcsUriReader
     import ValueReaders.PathReader
     
     //NB: Ficus now marshals the contents of loamstream.googlecloud.hail into a HailConfig instance.
