+++ src/test/scala/loamstream/googlecloud/HailConfigTest.scala
@@ -11,25 +11,13 @@ import loamstream.util.BashScript.Implicits._
  * Feb 24, 2017
  */
 final class HailConfigTest extends FunSuite {
-  test("jarFile") {
-    val jar = new URI("gs://foo/bar/baz")
-    val zip = new URI("gs://blerg/zerg/flerg")
-    val envName = "hail-0.2.18"
-
-    assert(HailConfig(jar, zip, envName).jarFile === "baz")
-  }
-
   test("fromConfig - good input, defaults used") {
-    val jar = new URI("gs://foo/bar/baz")
-    val zip = new URI("gs://blerg/zerg/flerg")
     val envName = "hail-0.2.18"
 
     val configString = s"""
       loamstream {
         googlecloud {
           hail {
-            jar = "$jar"
-            zip = "$zip"
             condaEnv = "$envName"
           }
         }
@@ -37,12 +25,10 @@ final class HailConfigTest extends FunSuite {
 
     val actual = HailConfig.fromConfig(ConfigFactory.parseString(configString)).get
 
-    assert(actual === HailConfig(jar, zip, envName, HailConfig.Defaults.scriptDir))
+    assert(actual === HailConfig(envName, HailConfig.Defaults.scriptDir))
   }
 
   test("fromConfig - good input, NO defaults used") {
-    val jar = new URI("gs://foo/bar/baz")
-    val zip = new URI("gs://blerg/zerg/flerg")
     val envName = "hail-0.2.18"
     val scriptDir = Paths.get("/foo/bar/baz")
 
@@ -50,8 +36,6 @@ final class HailConfigTest extends FunSuite {
       loamstream {
         googlecloud {
           hail {
-            jar = "$jar"
-            zip = "$zip"
             condaEnv = "$envName"
             scriptDir = "${scriptDir.render}"
           }
@@ -60,7 +44,7 @@ final class HailConfigTest extends FunSuite {
 
     val actual = HailConfig.fromConfig(ConfigFactory.parseString(configString)).get
 
-    assert(actual === HailConfig(jar, zip, envName, scriptDir))
+    assert(actual === HailConfig(envName, scriptDir))
   }
 
   test("fromConfig - bad input") {
